+++ src/users/user.entity.ts
@@ -1,5 +1,5 @@
 import { Entity, Column, PrimaryGeneratedColumn, CreateDateColumn, UpdateDateColumn, OneToMany } from 'typeorm';
-import { UsersToArticles } from './userstoarticles/userstoarticles.entity';
+import { UsersToArticles } from '../userstoarticles/userstoarticles.entity';
 
 @Entity()
 export class User {
