+++ src/users/users.controller.ts
@@ -14,9 +14,9 @@ import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/user.dto';
 import { ChangePasswordDto } from './dto/password.dto';
 import { UpdateResult } from 'typeorm';
-import { SaveArticleDto } from './articles/articles.dto';
-import { Article } from './articles/articles.entity';
-import { UsersToArticles } from './userstoarticles/userstoarticles.entity';
+import { SaveArticleDto } from '../articles/articles.dto';
+import { Article } from '../articles/articles.entity';
+import { UsersToArticles } from '../userstoarticles/userstoarticles.entity';
 import { UserSerialize } from './dto/users.serializer';
 import { AuthGuard } from '@nestjs/passport';
 import { ShareArticleDto } from './dto/shareArticle.dto';
